fix: EditTransaction layout

+++ src/components/TopPanel/TopPanel.js
@@ -4,6 +4,7 @@ import React, { useState } from 'react'
 
 import { ROUTER_PATH } from '../../Constant'
 import { atomState } from '../../Atoms'
+import { requestHandler } from '../../Services'
 import { useHistory } from 'react-router-dom'
 import { useRecoilValue } from 'recoil'
 
@@ -14,9 +15,14 @@ const TopPanel = () => {
   const [isDismissMenu, setIsDismissMenu] = useState(false)
 
   const onClickSetting = () => history.push(ROUTER_PATH.userSettings.path)
-  const onClickLogOut = () => {
-    window.localStorage.clear()
-    history.push(ROUTER_PATH.login.path)
+  const onClickLogOut = async () => {
+    try {
+      const { success } = await requestHandler('/logout', true, {}, 'post')
+      if (success) {
+        window.localStorage.clear()
+        history.push(ROUTER_PATH.login.path)
+      }
+    } catch (error) {}
   }
   return (
     <Container>
